♻️ refactor(test): inline PR 518 regression tests into test_async_filelock
Dedicated regression test files accumulate and fragment the test suite.
The two cases — sync context manager rejection and __del__ safety after
loop close — belong alongside the other async lock behavior tests.
